T350: 修告警趋势集成断言量纲（TrendRow.Value 是 float64）
CI 集成用例 TestT350ITTrendAndDistributionScope 报 expected int64(1) /
actual float64(1) —— repo 层 TrendRow.Value 统一是 float64，断言按 int64 写
是 testify 的严格类型判定，不是数据错。

+++ services/data-service/internal/repo/team_scope_t350_integration_test.go
@@ -217,8 +217,8 @@ func TestT350ITTrendAndDistributionScope(t *testing.T) {
 	require.NoError(t, err)
 	require.Len(t, alertAll, 1)
 	require.Len(t, alertB, 1)
-	assert.GreaterOrEqual(t, alertAll[0].Value, int64(2))
-	assert.Equal(t, int64(1), alertB[0].Value)
+	assert.GreaterOrEqual(t, alertAll[0].Value, 2.0)
+	assert.Equal(t, 1.0, alertB[0].Value)
 	assert.Empty(t, alertNone)
 
 	avgAll, err := r.PatientAvgWearMinutes(ctx, dateStr, model.ScopeAll())
